Extremely rickety keyboard navigation
This allows a user to navigate through comments with just the
keyboard. The bindings are as follows:

- p: previous unread comment
- n: next unread comment
- h: parent
- j: next comment in order (flattened hierarchy)
- k: previous comment in order (flattened hierarchy)
- l: first child

Only tested on comment pages for now, will likely break elsewhere.
No actions are bound.

The entire graph creation through jQuery traversal is again, rickety.
It does work though, although for now the assumption that comments
are laid out in order makes implementation simpler and not using
some of those fields for now.